Fixes the issue #166
Root cause is that in the list can be objects of different types (e.g. ParticipantClass and string for the items without class)
The default comparer cannot deal with that situation. It assumes objects of same type in the list. This fix is a custom comparer which sorts the strings at the end.